requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
gunicorn==22.0.0
pdf2docx==0.5.13
python-docx==1.2.0
google-genai==1.75.0
python-dotenv==1.2.2
